requirements.txt
ollama-python==0.1.2
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
scikit-learn==1.4.0
nltk==3.8.1
pytest==7.4.4
python-dotenv==1.0.0
pyyaml==6.0.1

# Streamlit UI dependencies
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4
